:truck: Move enums
