+++ scripts/enums.py
@@ -24,3 +24,33 @@ class StableDiffusionVersion(Enum):
             return StableDiffusionVersion.SDXL
 
         return StableDiffusionVersion.UNKNOWN
+
+
+class ControlModelType(Enum):
+    """
+    The type of Control Models (supported or not).
+    """
+
+    ControlNet = "ControlNet, Lvmin Zhang"
+    T2I_Adapter = "T2I_Adapter, Chong Mou"
+    T2I_StyleAdapter = "T2I_StyleAdapter, Chong Mou"
+    T2I_CoAdapter = "T2I_CoAdapter, Chong Mou"
+    MasaCtrl = "MasaCtrl, Mingdeng Cao"
+    GLIGEN = "GLIGEN, Yuheng Li"
+    AttentionInjection = "AttentionInjection, Lvmin Zhang"  # A simple attention injection written by Lvmin
+    StableSR = "StableSR, Jianyi Wang"
+    PromptDiffusion = "PromptDiffusion, Zhendong Wang"
+    ControlLoRA = "ControlLoRA, Wu Hecong"
+    ReVision = "ReVision, Stability"
+    IPAdapter = "IPAdapter, Hu Ye"
+    Controlllite = "Controlllite, Kohya"
+
+
+# Written by Lvmin
+class AutoMachine(Enum):
+    """
+    Lvmin's algorithm for Attention/AdaIn AutoMachine States.
+    """
+
+    Read = "Read"
+    Write = "Write"
